# swagger.yaml
openapi: 3.0.0
info:
  title: Chat Web API
  version: 1.0.0
  description: API documentation for the chat-web-api backend (all endpoints, public docs)
servers:
  - url: https://chat-web-api.onrender.com
  - url: http://localhost:8000
  - url: http://192.168.29.207:8000
paths:
...
  /api/users:
    get:
      summary: Get all users
      tags: [User]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of users
    put:
      summary: Update user profile
      tags: [User]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                profile_image:
                  type: string
                  format: binary
      responses:
        '200':
          description: Profile updated
  /api/users/logout:
    post:
      summary: Logout user (explicit)
      tags: [User]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout successful
  /api/users/search:
    get:
      summary: Search users
      tags: [User]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: q
          schema:
            type: string
          required: false
          description: Search query (legacy, use 'search' for new clients)
        - in: query
          name: search
          schema:
            type: string
          required: false
          description: Search query (preferred)
        - in: query
          name: limit
          schema:
            type: integer
          required: false
          description: Max results to return (default 20)
        - in: query
          name: offset
          schema:
            type: integer
          required: false
          description: Offset for pagination (default 0)
      responses:
        '200':
          description: List of matching users
  /api/users/change-password:
    post:
      summary: Change password
      tags: [User]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                oldPassword:
                  type: string
                newPassword:
                  type: string
      responses:
        '200':
          description: Password changed
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT